Added style to website to create flexible layout and generally make things look a bit nicer.
Added title to page, which has a link which resets all queryParams.

Added a feature to the pagination component where, instead of being disabled, the ellipse items open a dropup menu with links to all of the pages they're hiding.

Installed a mixin that scrolls the page back to the top whenever you click on a link.