added logic to map page that allows it to override the normal search query parameters, and then reset them when you navigate away.
Added loops to serializer that parse the `locations` data on each record, and returns the first set of coordinates it finds as `lat` and `lng` in the payload.  Added these as DS.attrs() for record.

Added leaflet function to display data locations on map, and added geo_bbox functionality to the model.

Updated style for search-result page to look better on small screens.