set host locale as a queryParam ('hl').
An oddity: right now, 'hl' is set to replace state, so it doesn't add a new item to browser history.  However, this means that if you change the language and then click the back button, it will both revert the language change and go back to the previous page, which may make the language change feel a bit less "sticky" than is desireable.  For the most part, I think it's fine though.